ПЕРЕПИСАЛ ВСЕ МЕТОДЫ ДО ТРАНСФЕРА

+++ src/main/java/com/example/bank/service/AccountService.java
@@ -6,6 +6,7 @@ import com.example.bank.mapper.AccountMapper;
 import com.example.bank.model.*;
 import com.example.bank.repository.AccountRepository;
 import com.example.bank.repository.UserRepository;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
@@ -23,6 +24,21 @@ public class AccountService {
     private final UserRepository userRepository;
     private static final Random random = new Random();
 
+    private User getCurrentUser() {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        return userRepository.findByUsername(username)
+                .orElseThrow(() -> new ResourceNotFoundException("User", "username", username));
+    }
+    private void assertAccountBelongsToCurrentUser(Account account) {
+        User currentUser = getCurrentUser();
+        if (!account.getUser().getUserId().equals(currentUser.getUserId())) {
+            throw new InvalidOperationException("Access denied: not your account");
+        }
+    }
+
+
+
+
     public AccountService(AccountRepository accountRepository, UserRepository userRepository) {
         this.accountRepository = accountRepository;
         this.userRepository = userRepository;
@@ -33,14 +49,21 @@ public class AccountService {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String username = authentication.getName();
 
+
+
         User user = userRepository.findByUsername(username)
                 .orElseThrow(() -> new ResourceNotFoundException("User", "username", username));
+
+
 
         if (accountDto.getBlocked() == null) {
             accountDto.setBlocked(false);
         }
 
         Account account = AccountMapper.toEntity(accountDto, user, generateUniqueId());
+        if (account.getBalance().compareTo(BigDecimal.ZERO) <= 0) {
+            account.setBalance(BigDecimal.ZERO);
+        }
         Account savedAccount = accountRepository.save(account);
 
         if (user.getMainAccount() == null) {
@@ -52,24 +75,53 @@ public class AccountService {
     }
 
     public Optional<Account> findById(Long id) {
-        return accountRepository.findById(id);
+        if (getCurrentUser().getRole() == Role.ADMIN) {
+            return accountRepository.findById(id);
+        } else {
+            throw new AccessDeniedException("Access denied: Current role is " + getCurrentUser().getRole());
+        }
     }
 
+
     public List<Account> findAll() {
-        return accountRepository.findAll();
+        if (getCurrentUser().getRole() == Role.ADMIN) {
+            return accountRepository.findAll();
+        } else {
+            throw new AccessDeniedException("Access denied: Current role is " + getCurrentUser().getRole());
+        }
     }
 
     public Account update(Account account) {
-        return accountRepository.save(account);
+        User user = getCurrentUser();
+
+        Optional<Account> existingAccountOpt = accountRepository.findById(account.getId());
+        if (existingAccountOpt.isEmpty()) {
+            throw new ResourceNotFoundException("Account", "id", account.getId());
+        }
+
+        Account existingAccount = existingAccountOpt.get();
+
+        if (user.getRole() == Role.ADMIN || existingAccount.getUser().getUserId().equals(user.getUserId())) {
+            return accountRepository.save(account);
+        } else {
+            throw new AccessDeniedException("Access denied: You can update only your own accounts");
+        }
     }
 
+
+
     public void setUserByUserId(Long userId, Account account) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new ResourceNotFoundException("User", "id", userId));
-        account.setUser(user);
+        if (getCurrentUser().getRole() == Role.ADMIN) {
+            User user = userRepository.findById(userId)
+                    .orElseThrow(() -> new ResourceNotFoundException("User", "id", userId));
+            account.setUser(user);
+        } else {
+            throw new AccessDeniedException("Access denied: Current role is " + getCurrentUser().getRole());
+        }
     }
 
     public void deposit(Long id, BigDecimal amount) {
+
         Account account = accountRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Account", "id", id));
 
@@ -84,6 +136,7 @@ public class AccountService {
     public void withdraw(Long id, BigDecimal amount) {
         Account account = accountRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Account", "id", id));
+        assertAccountBelongsToCurrentUser(account);
 
         if (amount.compareTo(BigDecimal.ZERO) <= 0) {
             throw new InvalidOperationException("Amount must be greater than zero");
@@ -101,7 +154,7 @@ public class AccountService {
         accountRepository.save(account);
     }
 
-    public long generateUniqueId() {
+    private long generateUniqueId() {
         long number;
         do {
             number = 10000000 + random.nextInt(90000000);
@@ -110,7 +163,23 @@ public class AccountService {
     }
 
     public void deleteById(Long id) {
-        accountRepository.deleteById(id);
+        Account account = accountRepository.findById(id).get();
+        User user = getCurrentUser();
+
+        Optional<Account> existingAccountOpt = accountRepository.findById(account.getId());
+        if (existingAccountOpt.isEmpty()) {
+            throw new ResourceNotFoundException("Account", "id", account.getId());
+        }
+
+        Account existingAccount = existingAccountOpt.get();
+
+        if ((user.getRole() == Role.ADMIN || existingAccount.getUser().getUserId().equals(user.getUserId())) && account.getBalance().compareTo(BigDecimal.ZERO) == 0) {
+            accountRepository.deleteById(id);
+        } else {
+            throw new AccessDeniedException("Access denied: You can update only your own accounts");
+        }
+
+
     }
 
     @Transactional
